add lease_term and move_in_date to payload

+++ src/app/fetch.js
@@ -1,8 +1,10 @@
 import { ROUTES } from 'app/constants';
+import { MOCKY } from 'app/api';
 
 // fetch middleware handles invalid token/expiry
 export default async (...args) => {
     const res = await fetch(...args);
+    if ( MOCKY ) return Promise.reject();
     if (res.headers.get('error_type') === 'TokenExpiredError') {
         if (localStorage.access_token) {
             localStorage.clear();
